NEXLOAD2: V1.3 prototype of new fields + prototype binary
WARNING:

I think the loader itself needs now core 3.0.5+ to work correctly
(loading screens setup), but it does *NOT* check if it is running on new
cores.

not fully tested yet (need to update also tools to create the files
easily, binary patching was used to test some of it)

And the fields definitions are still subject to change, based on
eventual feedback.

Check https://wiki.specnext.dev/NEX_file_format for current status.